+++ src/realm/object-store/sync/sync_manager.cpp
@@ -77,9 +77,6 @@ void SyncManager::configure(std::shared_ptr<app::App> app, const std::string& sy
 
         // Set up the metadata manager, and perform initial loading/purging work.
         if (m_metadata_manager || m_config.metadata_mode == MetadataMode::NoMetadata) {
-            // No metadata means we use a new client uuid each time
-            if (!m_metadata_manager)
-                m_client_uuid = util::uuid_string();
             return;
         }
 
@@ -99,7 +96,6 @@ void SyncManager::configure(std::shared_ptr<app::App> app, const std::string& sy
         }
 
         REALM_ASSERT(m_metadata_manager);
-        m_client_uuid = m_metadata_manager->client_uuid();
 
         // Perform our "on next startup" actions such as deleting Realm files
         // which we couldn't delete immediately due to them being in use
@@ -210,7 +206,6 @@ void SyncManager::reset_for_testing()
     {
         util::CheckedLockGuard lock(m_file_system_mutex);
         m_metadata_manager = nullptr;
-        m_client_uuid = util::none;
     }
 
     {
@@ -699,12 +694,6 @@ SyncClient& SyncManager::get_sync_client() const
 std::unique_ptr<SyncClient> SyncManager::create_sync_client() const
 {
     return std::make_unique<SyncClient>(make_logger(), m_config, weak_from_this());
-}
-
-std::string SyncManager::client_uuid() const
-{
-    REALM_ASSERT(m_client_uuid);
-    return *m_client_uuid;
 }
 
 util::Optional<SyncAppMetadata> SyncManager::app_metadata() const